+++ jeap-messaging-contract-annotation-processor/src/main/java/ch/admin/bit/jeap/messaging/annotations/processor/util/TemplateMessageCollector.java
@@ -10,10 +10,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.stream.Stream;
 
 public class TemplateMessageCollector {
@@ -25,8 +22,8 @@ public class TemplateMessageCollector {
     private static final String EVENT_NAME_KEY = "eventName";
     private static final String TOPIC_NAME_KEY = "topicName";
 
-    public Map<String, List<String>> collectTemplateMessages(String templatePath, Messager messager) {
-        Map<String, List<String>> templateMessageMap = new HashMap<>();
+    public Map<String, Set<String>> collectTemplateMessages(String templatePath, Messager messager) {
+        Map<String, Set<String>> templateMessageMap = new HashMap<>();
 
         try (Stream<Path> paths = Files.walk(Paths.get(templatePath))) {
             paths.filter(Files::isRegularFile)
@@ -39,7 +36,7 @@ public class TemplateMessageCollector {
         return templateMessageMap;
     }
 
-    private void processFile(Path path, Map<String, List<String>> templateMessageMap, Messager messager) {
+    private void processFile(Path path, Map<String, Set<String>> templateMessageMap, Messager messager) {
         try {
             String content = Files.readString(path);
             JsonObject jsonObject = JsonParser.parseString(content).getAsJsonObject();
@@ -51,14 +48,14 @@ public class TemplateMessageCollector {
         }
     }
 
-    private void addTemplateMessagesToMap(Map<String, List<String>> templateMessageMap, JsonArray templateMessageJsonArray) {
+    private void addTemplateMessagesToMap(Map<String, Set<String>> templateMessageMap, JsonArray templateMessageJsonArray) {
         if (templateMessageJsonArray != null) {
             for (int i = 0; i < templateMessageJsonArray.size(); i++) {
                 JsonObject contract = templateMessageJsonArray.get(i).getAsJsonObject();
                 String name = contract.has(MESSAGE_NAME_KEY) ? contract.get(MESSAGE_NAME_KEY).getAsString() : contract.get(EVENT_NAME_KEY).getAsString();
                 String topicName = contract.get(TOPIC_NAME_KEY).getAsString();
 
-                templateMessageMap.computeIfAbsent(name, k -> new ArrayList<>()).add(topicName);
+                templateMessageMap.computeIfAbsent(name, k -> new TreeSet<>()).add(topicName);
             }
         }
     }