+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.30.3] - 2025-03-28
+
+### Fixed
+
+- Message Contract Annotation Processor: Avoid creating duplicate contracts for the same message type / topic combination
+  when generating contracts from templates. 
+
 ## [8.30.2] - 2025-03-26
 
 ### Fixed